Falcon: Added simple MXML feature tests for <Boolean>, <int>, <uint>, <Number>, <String>, <Class> and generic instance tags such as <Sprite>. They are not yet part of 'ant tsets'.
These tests currently create an MXML file, compile it to a SWF with MXMLC, and check that it compiles cleanly. Soon I'll make the SWF run in Flash Player and check that the runtime asserts don't fire, so that we know that codegen is correct.

Such "feature tests" are intended to test one specific feature of MXML all the way through compilation and execution. By contrast, "unit tests" test only a particular class or subsystem, such as testing the MXML lexer to create MXML tokens, or testing how the MXML parser creates particular kinds of MXML nodes. Furthermore, "functional tests" test real-world situations involving multiple features working together. There are three distinct directories for these various kinds of Falcon tests.

Feature tests do not need to be as exhaustive as the unit tests. For example, suppose we have a complete set of parser tests for the <int> tag and know that ultimately an MXMLLiteralNode gets created with an appropriate int value in each case. Then we do not need to have feature tests for things like integer tags with whitspace around the integer text, or hex integer text, or empty int tags, etc. We can just do a few feature tests that make sure that MXMLLiteralNodes with int values codegen integer values in the VM.


git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/falcon/trunk@1415883 13f79535-47bb-0310-9956-ffa450edef68