Added basic feature tests for the <Object> tag.
One of the tests is marked @Ignore because it doesn't pass. Falcon understand <Object> tags that have properties, like <Object a="1" b="2"/>, but it doesn't understand primitive ones like <Object>abc</Object>.

One of the tests that now passes uncovered a bug in MXMLArrayNode which I fixed. The problem was that a dynamic property like 'a' in <Object a="1" b="2"/> has no definition, because it isn't declared anywhere.

Feature tests that fail due to compilation problems now show the problems.

Enhanced the getMXML() method of MXMLInstanceTagTestsBase so that you can inject more code into the <Script> block, such as test functions that you want to use in asserts.



git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/falcon/trunk@1418562 13f79535-47bb-0310-9956-ffa450edef68