Falcon: Made it easy to inspect the ABC produced by any compilation unit.
I've added a toString() method to ABCBytesRequestResult that produces an ABC dump using ABCDumpVisitor. You can set a breakpoint in a compilation unit's handleABCBytesRequest() (for example, in MXMLCompilationUnit if you are compiling MXML) and just look at the ABCBytesRequestResult in the debugger to see the dump. This is similar to how you can see entire ASTs and scopes in the debugger.

git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/falcon/trunk@1418588 13f79535-47bb-0310-9956-ffa450edef68