Falcon feature tests: Added a feature test for the <Metadata> tag.
It appears that Falcon doesn't automatically keep such metadata; you have to use -keep-as3-metadata if it isn't one of the types like Event and Style that are kept by default. This is probably a bug.

Made it possible to pass additional compilation options like -keep-as3-metadata when compiling feature tests.

git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/falcon/trunk@1425170 13f79535-47bb-0310-9956-ffa450edef68