Falcon tests: Added some functional tests for the <XMLList> tag.
Added one test and fixed a bug related to <XML format="xml"> tag, which produces an XMLNode. However, there is an XML namespace issue preventing the test from passing (because when the tag gets converted into an XML string Falcon is adding empty namespace attributes that the old compiler didn't) so this test is marked @Ignore.

git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/falcon/trunk@1425203 13f79535-47bb-0310-9956-ffa450edef68