Falcon: Deleted compiler.tests/unittest.properties.
Developers working on Falcon may create this file with their paths for FLEX_HOME, AIR_HOME, PLAYERGLOBAL_HOME, and FLASHPLAYER_DEBUGGER but should not check it in because other developers like to set these paths via environment variables. There is already a template-unittest.properties which you can copy and customize if you want.

git-svn-id: https://svn.apache.org/repos/asf/flex/falcon/trunk@1432328 13f79535-47bb-0310-9956-ffa450edef68