Falcon: Added MXML parser unit tests that create a MXMLPropertySpecifierNode for properties of type Boolean, int, uint, Number, and String, using property attributes, property child tags with text, and property child tags with instance tags.
To keep these tests single-file for simplicity, they use a <Component> tag to define a class with a property of a particular type. Then they create an instance tag for that class on which they set the property in the various ways.

Fixed a bug uncovered by this work, in which an MXML tag referring to a class defined by a <Component> tag didn't resolve correctly.

Eliminated FAILURE messages in Ant coming from unit-test base classes that don't have any JUnit tests of their own.


git-svn-id: https://svn.apache.org/repos/asf/flex/falcon/trunk@1435423 13f79535-47bb-0310-9956-ffa450edef68