FlexJS MXML in FalconJx
This is the implementation of the first step towards full FlexJS support in FalconJx. I started with the MXML because that needed the most work. In this version there are some workarounds, particularly when dealing with 'scripts' (getter, event handlers), but those will be addressed when the JS emitter for FlexJS is implemented. For now, have a look at this contribution and provide some feedback, please.