Remove generic .gitignore
Removed the generic .gitignore in order to replace it with project specific .gitignore files that contain ONLY the files that need to be ignored by that particular project. The generic .gitignore is too all inclusive to be "safe to use" as it might exclude files that actually need to be in/added to the project.

Signed-off-by: Erik de Bruin <erik@ixsoftware.nl>