[FalconJX] full AS project test capability
Added the ability to test full AS projects to TestBase.java. Since some functionalities of AS only become apparent when compiling projects consisting of 2 or more classes, this ability is needed as part of the full test suite.

Note: this commit adds 2 libraries to the compiler.jx.tests build path: '/compiler.jx/lib/commons-io.jar' and '/compiler.jx/lib/google/closure-compiler/compiler.jar', both of which are already part of the compiler.jx project.
Signed-off-by: Erik de Bruin <erik@ixsoftware.nl>