Deleted obsolete MXML tokens related to databinding.
These were TOKEN_DATABINDING, TOKEN_DATABINDING_START,
and TOKEN_DATABINDING_END.

I also deleted MXMLDatabindingData and MXMLDatabindingValue
and their corresponding interfaces.

These were vestiges of an old Adobe attempt to unify
the parsing of MXML and the parsing of E4X in AS.
This turned out to be a bad idea because of the
peculiarities of E4X.

A databinding expresssion in MXML cannot be considered a
single "unit" of MXML (i.e., one MXMLUnitData).
For example, a databinding expression inside a property tag
can span multiple MXML units representing text, comments,
CDATA, etc.

As a result, databinding parsing is not part of the syntactic
parsing of MXML that creates the MXML tokens and the MXMLData.
It is part of the higher-level semantic analysis that builds
the MXML AST.