[FalconJX] updated test files
The FalconJX test files were updated to reflect the change from 'self' to 'this' references and the removal of 'var self = this;'.

The built-in test class was renamed from 'A' to 'FalconTest_A' to prevent name clashes. The built-in method 'a' was renamed to 'falconTest_a' as well.

A test file 'TestFlexJSFile' was added, containing tests that use an entire external AS class file as input.

Signed-off-by: Erik de Bruin <erik@ixsoftware.nl>