Removed support for MXML "tag blobs".
This was a old performance optimization from when CodeModel
and the old compiler were separate parts of FlashBuilder.
It allowed CodeModel's MXML lexer/parser to skip over
the contents of an <fx:Private> tag.

But allowing an <fx:Private> tag to have malformed content
is antithetical to the idea of MXML being XML.
Falcon is both a compiler and a code intelligence engine,
and it must check the lexical structure of the entire
MXML file.

This gets rid of
- PrivateTagDetector and its interface ITagAggregateDector
- TOKEN_MXML_BLOB
- MXMLTagBlobData and its interface IMXMLTagBlobData