Fix: EXTERNC - IMPORT NOT EMITTED
Imports relative to some function return type
are now emitted only if either the static member or the
return type class are not excluded.

This includes that:

 Given a function A():ns.ReturnType
  And A is not excluded
  And ns.ReturnType is not excluded
Then the import of ns.ReturnType will occur.

 Given a function A():ns.ReturnType
  And A is not excluded
  And ns.ReturnType is excluded
Then the import of ns.ReturnType will NOT occur.

 Given a function A():ns.ReturnType
  And A is excluded
  And ns.ReturnType is not excluded
Then the import of ns.ReturnType will NOT occur.

 Given a function A():ns.ReturnType
  And B():ns.ReturnType
  And B is excluded
  And ns.ReturnType is not excluded
 Then the import of ns.ReturnType still occur.