compiler: added allow-import-aliases command line option
I previously added support for import aliases to the ActionScript parser. However, they were always enabled, even when using the Royale compiler for IDE code intelligence with a different SDK that doesn't support them. To ensure that an appropriate syntax error will appear for other SDKs, and to ensure that new syntax is opt-in, import aliases are disabled by default, and can be enabled by using this new option.