compiler-jx: js-default-initializers is true by default in the compiler
As previously discussed, this is a better default in the general case because it preserves ActionScript's existing behavior. Framework, library, and app authors can choose to turn it off as an optimization, with the knowledge of the side effects that it entails, if desired. I preemptively set js-default-initializers to false in the various library projects in royale-asjs before this commit because that's what the Royale framework authors want.