compiler-jx: most console output during compilation is hidden and can be re-enabled with the -verbose option
This change adds a noticeable improvement to compile times. Compiling TourDeJewel takes 2 seconds less on my machine. Additionally, the classic Flex SDK and ASC 2.0 don't output all of that stuff, so it makes Royale more consistent with the rest of the ecosystem. Plus, the reduced noise makes errors and warnings seem more important, as they should be.