compiler-jx: added inline-constants compiler option for JS builds to optionally inline the values of primitive constants (like Number, String, Boolean, int, and uint) instead of referencing them.
Defaults to false. It may make sense to switch this to true in the future, but it's not stable yet.

Does not yet skip unnecessary goog.require() calls, even though the classes may no longer be referenced. Figuring out how to do this should reduce the size of release builds.