Definitions containing [JSModule] with a custom module name are no longer required to use strict camelCase naming scheme.
Previously, if a module was named my-module, the definition was required to be named myModule. Now, it can be named anything that is a valid identifier, and the compiler will output that identifier for the require() call. Example:

[JSModule(name="my-module")] public class MyCustomModule {}

Becomes:

var MyCustomModule = require('my-module');

Additionally, if the definition is in a package, a custom module name is not strictly required anymore. If no custom module name is specified, the first part of the package name is used as the module name. For instance, com.example.MyClass without a custom module name automatically requires 'com'.