COMPJSC: fix issue where .js files from one .swc file on the library-path were not copied to the new .swc file
The bytecode was being included, but not the corresponding .js file. Both are required, and this broke the app compiler, which couldn't figure out where to find the .js file. To exclude both the bytecode and .js file, then the .swc file should be added to the external-library-path instead of the library-path.