AccessorEmitter: always use nocollapse for static accessors (closes #200)
Closure compiler seems to correctly rename accessors when they are on the prototype, but it only renames references when accessors are static, so nocollapse is needed as a workaround. We might be able to remove this in the future, if Closure ever fixes the issue. For now, since this code is not dynamic (we allow dynamic code to break when we don't prevent renaming), this nocollapse workaround is necessary to allow setting prevent-rename compiler options to false without breaking non-dynamic code.