BinaryOperatorEmitter: Fix setting Date properties inside chained assignment (closes #212)
Previously, something like this would result in NaN for hours and minutes: d.hours = d.minutes = d.seconds = xyz

This is because setSeconds/setMinutes and other methods return void. The fix is to use a comma operator and call getSeconds/getMinutes so that the value propagates. The double function call should be wrapped in parentheses to avoid confusion with commas in other contexts, such as function calls.

The code in the example above becomes: d.setHours((d.setMinutes((d.setSeconds(xyz), d.getSeconds())), d.getMinutes()));

Notice that the setHours() call doesn't include a matching getHours() call because it's not necessary to propagate the value further.