New compiler option --infer-types to enable type inference for variables, constants, parameters, and return types
If no type is specified, but the variable is assigned in the declaration, the type is inferred automatically from the right hand side of the assignment. For functions, the return type is inferred from the types of return statements (or lack thereof), resolving a common base type, if necessary.

These types are used not only for compile time problem checking, but also in the generated ABC bytecode and the generated JS, the same as if the type was explicitly declared. This ensures maximum performance (and best optimization by Closure Compiler), and maximum productivity in editors and IDEs that use types for code intelligence.

Disabled by default and must opt-in to use (we could consider making it opt-out in the future, though).