MXMLRoyaleEmitter: When writing MXML properties to JS, escape all backslash characters
For instance, \n is not a new line in MXML strings. It becomes \\n instead. To insert a real \n new line that won't be modified, use &#xA; in MXML instead. This is how the Flex SDK compiler behaves, so the Royale compiler now more closely matches Flex.

Also cleaned up handling of carriage return and tab characters to match the Flex SDK compiler.