New [RoyaleCallableInstances] meta may be added to a typedef class to allow instances of the class to be called like a function, to support the jQuery $ symbol both being callable and having members
This basically reverts commit 1152253d28ce140f5e1339e525645ec9e64424fe, which made the compiler more lenient when calling a variable or accessor to support jQuery. That change unfortunately moved errors on most objects that are not callable from compile-time to run-time, which is not ideal. This new metadata allows us to go back to the same strictness as the Flex SDK compiler for most code, with the option to opt into the extra lenience on specific classes.

This metadata does not currently allow defining a function signature, including checking parameter and return types, but that wasn't possible before either. It could potentially be added through parameters on the metadata in the future, though.

(closes #112)