js-include-asset compiler options
Works with both mxmlc and compc. Allows specifying arbitary files that will be included in the SWC and copied to the output directory of a compiled app. These files may be loaded at run-time from AS3, or they can be referenced from .js or .css files included with js-include-script or js-include-css.