FunctionScope: fix exception where a FunctionNode has been garbage collected, and a new FunctionNode gets re-attached to the existing FunctionScope scope
Similar to FunctionNode's discardFunctionBody() method, we need to clear out any old definitions that will no longer be valid because the new function body (even if it has the same contents) will parse new VariableNodes and things for local variables and create new definitions to replace the old ones.

This issue is definitely an edge case, but I've hit it more than once over the years. It seems to be uncommon for FunctionNodes to get garbage collected, but when they do, an assertion failure in parseFunctionBody() gets triggered. In particular, this was causing the ASDoc JSON generation to fail intermittently (but relatively often, since it parses so many classes). It was originally doing that silently (so files weren't generated, but the ASDoc tool reported success), but I fixed that in previous commits.

In the past, I think that I encountered this same issue when implementing the -watch compiler option. When watching, the compiler continues to run after the initial compilation and then recompiles on source path file changes. Running indefinitely made it more likely for the GC to eventually run and FunctionNodes to be replaced. I found a workaround for that at the time (I think using workspace.startBuilding() and  workspace.doneBuilding()), but that didn't seem to work here. That makes me think that it really was a workaround and only helped, but didn't solve the issue.

Removing definitions like discardFunctionBody() when the FunctionScope detects that function body hasn't been parsed yet seems like it will be much more reliable.

The reason that FunctionNodes can be GCed is that a weak reference to the AST is kept by ASCompilationUnit's syntax tree request. At first, it holds a hard reference, but at some point later in the compilation process, it is deemed safe to be GCed. From that point, it is demoted to a weak reference. This appears to be by design in the code that was donated to us by Adobe. This process of demoting from strong to weak reference technically allows the FunctionNode (technically, I think it's the FileNode that contains the FunctionNode, which is usually a method on a ClassNode) to remain accessible for a while, and I think that's part of why it was so difficult to reproduce this issue every time I've encountered it. It's hard to determine when GC is run, and if the weak reference is grabbed and stored somewhere else as a strong reference, it won't be GCed until that other strong reference is removed. This time around, I added a lot of System.gc() calls, which seemed to help with reproducing more consistently.

I'm not sure if FunctionScope is the best place to put this code. There may be a way to put it into FunctionNode instead, which would keep the quirks of FunctionNodes being able to defer parsing bodies or to discard bodies a bit more encapsulated. However, even in FunctionScope, that's still pretty specific to functions.