debugger: add and use javac.src, javac.deprecation, and javac.debug properties, like other Java modules
In particular, we want to target JDK 11, instead of the version of the current JDK being used. A value wasn't being passed for the source and target values to Ant's <javac> task, which meant it used the current JDKs version by default. As a result, our previous release had fdb.jar require JDK 17, since that seems to have been the version used to build the release.