JSIncludeScript, JSIncludeCSS, and JSIncludeAsset metadata
These are similar to the -js-include-script, -js-include-css, and -js-include-asset compiler options, except that they associate these files with the specific package-level definition. So, if they are compiled into a SWC, the files are copied, and <script>/<link> tags are added, only if that definition is actually used. The compiler options do that if any definition in a SWC is used.

Works similarly to the classic [Embed] metadata with a source attribute to specify a file path that is either absolute or relative to the current .as or .mxml source file.