ASParser: partial fixes for null conditional (a?.b) operator
In particular, it wasn't handling deep nesting of null conditionals, method calls, or dynamic access properly. For method calls and dynamic access, the generated ABC and JS could attempt to perform the action on null, instead of skipping the action like it should have.

This is because ASParser was detecting only the nested ?. or . member accesses, but not the () or [] parts before transforming the null conditional into a ternary operator.

ASParser now parses what follows the null conditional operator differently than the regular member access operator. In particular, it includes a "name expression", followed by an optional function call, dynamic access, regular member access, namespace access, or additional null conditionals.

Includes a ton of new tests, not only for the bugs that this fixes, but also some things that would fail with other buggy implementations that were tried before coming up with the current solution.

This much better than the previous implementation (which is why I'm committing before fixing all issues), but still not perfect. If the method call or dynamic access doesn't immediately follow a null conditional, but follows a member access that follows the null conditional, the action can still be placed outside the ternary, which means it may end up being performed on null. An easy workaround for AS3 developers is just to convert all regular member access following null conditionals into null conditionals too.