Arrow functions
- Omits function keyword in favor of => arrow symbol between signature and body
- Parentheses are optional for a single parameter that is an identifier with no type or default value
- When parentheses are omitted, the parameter's type is not inferred and the compiler reports a warning
- Braces are optional for a single expression (no statements)
- If braces are omitted, the single expression becomes the return value
- If braces are omitted, the return type is optional and will be inferred (without enabling other type inference)
- 'this' is bound to the same value as the containing scope

Does not currently use JS arrow functions in the implementation. Nor does it use
Function.prototype.bind() because that doesn't exist in SWF either. Both SWF and
JS share a custom implementation that rewrites the arrow function like this:

```
function(context:*, func:Function):Function {
  return function(...rest):* {
    return func.apply(context, rest);
  };
}(this, function(/* params */)/*: return_type */ {
  /* body */
});
```

That doesn't mean that we can't use real JS arrow functions in the future. Just
starting out with something that all works for SWF, and we can refine them as
needed later.