null conditional operator with dynamic access
Example: a?.['b']

This new syntax enables a combination of the null conditional operator (a?.b) and dynamic access (a['b']). In the example above, if a is nullish, the operator will return null instead of trying to access the field named 'b'.

With a?.b, the name of b must be known at compile-time. With a?.['b'], the field name b can be generated at run-time.