MXMLClassDefinitionNode: allow Model tag as child of root when using MXML 2006 dialect
Moved Model child parsing from MXMLDocumentNode into MXMLClassDefinitionNode instead. This allows mx:Model to be declared in custom mx:Component tags too. This was not allowed before, but should have been. Improves consistency with Flex SDK compiler.

MXML 2009 does not allow Models to be declared as child of the root, giving an error about not implementing IUIComponent. This is true even if the root tag is fx:Object, which wouldn't have any requirement for IUIComponent. So, with MXML 2009, Models should go always into fx:Declarations, which we were not requiring before, but should have been. Further improves consistency with Flex SDK compiler.