JSCSSCompilationSession: improved distinction between components and HTML elements when outputting CSS
Followup to commit 3f09ebc41206a2578745e14ba6d705b7c4b3b6a2, which could incorrectly treat components in the default namespace as HTML elements.

Now checks if the element name is in the known htmlElementNames list (which needs to be updated, actually), but also checks if the resolved qname contains a "." character, which would indicate an AS3/MXML class in a package, which is definitely not an HTML element. I added that extra check to avoid the possibility of a class name like Button possibly matching an element like <button>. I also removed the namespace prefix check I was using in that commit to allow the XHTML namespace to optionally have a prefix in the future.