Function type expressions: optionally specify a required function signature for variables that can store functions, and check that signature at compile-time
Compiles down to the Function type, so there are no additional checks at run-time.

var f:(s:String, n?:Number, ...rest)=>void;

Similar syntax to real functions. Parameters and return type are separated by the => token. Optional parameters use the ? token instead of specifying a default value with = and a constant. This allows functions to have different default values.

Includes various checks for valid use of the signature.

- Assigning another variable that has a function type expression.
- Assigning a real function to the variable (including arrow functions).
- Calling the variable in a function call expression.

If an assignment includes a nested function type expression, such as a parameter or return type, that is checked too.